chore: finish deleting google-map-related modues
