+++ config/runtime.exs
@@ -200,9 +200,6 @@ if config_env() == :prod do
 end
 
 config :dotcom, LocationService,
-  google_api_key: System.get_env("GOOGLE_API_KEY"),
-  google_client_id: System.get_env("GOOGLE_MAPS_CLIENT_ID") || "",
-  google_signing_key: System.get_env("GOOGLE_MAPS_SIGNING_KEY") || "",
   aws_index_prefix: System.get_env("AWS_PLACE_INDEX_PREFIX", "dotcom-prod")
 
 config :dotcom, DotcomWeb.ViewHelpers,
