+++ lib/dotcom_web/templates/partial/_hidden_icons.html.eex
@@ -42,9 +42,6 @@
   <div id="icon-feature-alert">
     <%=  svg_icon(%SvgIcon{icon: :alert, class: "icon-small-inline"}) %>
   </div>
-  <div id="powered-by-google-logo">
-    <img src="<%= static_url(@conn, "/images/powered_by_google_on_white.png") %>"></img>
-  </div>
   <div id="icon-feature-map-pin">
     <%=  svg("icon-map-pin.svg") %>
   </div>
